+++ Assets/_Application/Configuration/FunctionalitiesPane.cs
@@ -1,5 +1,5 @@
-using System.Collections;
 using System.Collections.Generic;
+using Netherlands3D.Twin.Configuration.UI;
 using Netherlands3D.Twin.Functionalities;
 using TMPro;
 using UnityEngine;
