+++ Assets/_Application/Configuration/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7152e838989184b4898223158d88fd87
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
